fullstackdeveloper/python/fastapi/twillosms/requirements.txt
fastapi
uvicorn
pydantic
pyotp
twilio
python-jose
passlib
python-dotenv
bcrypt